feat : BoardFactory 추가 구현
  > BoardFactory 클래스
   - Board 인스턴스를 리턴하는 팩토리 메소드들
